add empty mac printer class
